per_device_train_batch_size < 1.

+++ hf/utils.py
@@ -50,7 +50,7 @@ def get_synthetic_data_device_iterator(config, tokenizer, mesh):
     # process that handles all runtime devices.
     assert xr.world_size() == 1
     num_devices = xr.global_runtime_device_count()
-    global_batch_size = config.per_device_train_batch_size * num_devices
+    global_batch_size = int(config.per_device_train_batch_size * num_devices)
     data = {
         "chosen_input_ids": torch.randint(tokenizer.vocab_size, (global_batch_size, config.max_length), dtype=torch.int64),
         "chosen_attention_mask": torch.ones(global_batch_size * 2, config.max_length, dtype=torch.int64),
